.github/scripts/requirements.txt
pandas==2.1.4
PyGithub==2.1.1
python-dateutil==2.8.2
pytz>=2024.1 